requirements.txt
git+https://github.com/KurimuzonAkuma/pyrogram
asyncpg
yt-dlp
tgcrypto
validators
py-tgcalls
python-dotenv
uvloop
click
